Updates to the Unit Switcher Pie Menu
Added Center & Ground Script, it centers a meshes vertices, freezes the
mesh, moves the center to the bottom, and moves the mesh to the 0 axis.

Updated the thicken .35 acetate macro, I can't decide if I want it to
assign a material or not.

Added an open in default application button to the right click menu of
an image in the shader tree.  This is super usefull.